final_projecttwo/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
Flask-APScheduler==1.11.0
Flask-Dance==5.0.0
Python-Dotenv
